+++ dashboard/src/components/forms/attributes/Contagion/HIV/sub/DueToBloodPrd.tsx
@@ -1,13 +1,13 @@
-// Has a family member infected with HIV due to blood products
+// Infected with HIV due to blood product administration
 
 import { useState, useCallback, useEffect } from 'react';
 import { useNavigationType } from 'react-router-dom';
 import { Checkbox } from '@chakra-ui/react';
 
 import { useRecoilState, useRecoilValue } from 'recoil';
-import { currentDateAtom, householdAtom } from '../../../../../state';
+import { currentDateAtom, householdAtom } from '../../../../../../state';
 
-export const HasFamily = ({ personName }: { personName: string }) => {
+export const DueToBloodPrd = ({ personName }: { personName: string }) => {
   const navigationType = useNavigationType();
   const currentDate = useRecoilValue(currentDateAtom);
 
@@ -18,11 +18,11 @@ export const HasFamily = ({ personName }: { personName: string }) => {
   const onChange = useCallback((event: React.ChangeEvent<HTMLInputElement>) => {
     const newHousehold = { ...household };
     if (event.target.checked) {
-      newHousehold.世帯員[personName].家族に血液製剤によるHIV感染者がいる = {
+      newHousehold.世帯員[personName].血液製剤の投与によってHIVに感染した = {
         [currentDate]: true,
       };
     } else {
-      newHousehold.世帯員[personName].家族に血液製剤によるHIV感染者がいる = {
+      newHousehold.世帯員[personName].血液製剤の投与によってHIVに感染した = {
         [currentDate]: false,
       };
     }
@@ -33,9 +33,9 @@ export const HasFamily = ({ personName }: { personName: string }) => {
 
   // stored states set checkbox when page transition
   useEffect(() => {
-    const hasFamilyObj =
-      household.世帯員[personName].家族に血液製剤によるHIV感染者がいる;
-    setIsChecked(hasFamilyObj && hasFamilyObj[currentDate]);
+    const dueToBloodObj =
+      household.世帯員[personName].血液製剤の投与によってHIVに感染した;
+    setIsChecked(dueToBloodObj && dueToBloodObj[currentDate]);
   }, [navigationType]);
 
   return (
@@ -46,7 +46,7 @@ export const HasFamily = ({ personName }: { personName: string }) => {
         colorScheme="cyan"
         mb={2}
       >
-        家族に血液製剤によるHIV感染者がいる
+        血液製剤の投与によってHIVに感染した
       </Checkbox>
       <br />
     </>
