+++ dashboard/src/components/forms/attributes/Contagion/HepatitisC/sub/BloodProduct.tsx
@@ -1,13 +1,13 @@
-// Infected with HIV due to blood product administration
+// 血液製剤の投与によってC型肝炎ウイルスに感染した
 
 import { useState, useCallback, useEffect } from 'react';
 import { useNavigationType } from 'react-router-dom';
 import { Checkbox } from '@chakra-ui/react';
 
 import { useRecoilState, useRecoilValue } from 'recoil';
-import { currentDateAtom, householdAtom } from '../../../../../state';
+import { currentDateAtom, householdAtom } from '../../../../../../state';
 
-export const DueToBloodPrd = ({ personName }: { personName: string }) => {
+export const BloodProduct = ({ personName }: { personName: string }) => {
   const navigationType = useNavigationType();
   const currentDate = useRecoilValue(currentDateAtom);
 
@@ -18,11 +18,15 @@ export const DueToBloodPrd = ({ personName }: { personName: string }) => {
   const onChange = useCallback((event: React.ChangeEvent<HTMLInputElement>) => {
     const newHousehold = { ...household };
     if (event.target.checked) {
-      newHousehold.世帯員[personName].血液製剤の投与によってHIVに感染した = {
+      newHousehold.世帯員[
+        personName
+      ].血液製剤の投与によってC型肝炎ウイルスに感染した = {
         [currentDate]: true,
       };
     } else {
-      newHousehold.世帯員[personName].血液製剤の投与によってHIVに感染した = {
+      newHousehold.世帯員[
+        personName
+      ].血液製剤の投与によってC型肝炎ウイルスに感染した = {
         [currentDate]: false,
       };
     }
@@ -33,9 +37,10 @@ export const DueToBloodPrd = ({ personName }: { personName: string }) => {
 
   // stored states set checkbox when page transition
   useEffect(() => {
-    const dueToBloodObj =
-      household.世帯員[personName].血液製剤の投与によってHIVに感染した;
-    setIsChecked(dueToBloodObj && dueToBloodObj[currentDate]);
+    const bloodProductObj =
+      household.世帯員[personName]
+        .血液製剤の投与によってC型肝炎ウイルスに感染した;
+    setIsChecked(bloodProductObj && bloodProductObj[currentDate]);
   }, [navigationType]);
 
   return (
@@ -46,7 +51,7 @@ export const DueToBloodPrd = ({ personName }: { personName: string }) => {
         colorScheme="cyan"
         mb={2}
       >
-        血液製剤の投与によってHIVに感染した
+        血液製剤の投与によってC型肝炎ウイルスに感染した
       </Checkbox>
       <br />
     </>
